backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.12
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
telethon==1.28.5
python-dotenv==1.0.0